requirements.txt
cffi==1.14.5
click==4.0
cryptography==3.3.2
decorator==3.4.2
Django==1.8
django-allauth==0.19.1
django-crispy-forms==1.4.0
django-leaflet==0.16.0
enum34==1.0.4
future==0.18.2
geocoder==1.38.1
geopy==1.10.0
ipaddress==1.0.23
mysqlclient==1.4.6
ndg-httpsclient==0.3.3
oauthlib==0.7.2
pyasn1==0.1.7
pycparser==2.10
pyOpenSSL==20.0.1
python-decouple==3.4
python-openid==2.2.5
ratelim==0.1.6
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
style==1.1.0
update==0.0.1
